prettier, rearrange function order to match PR

+++ contracts/test/ICErc20.sol
@@ -3,12 +3,19 @@
 pragma solidity ^0.7.3;
 
 interface ICErc20 {
-    function mint(uint mintAmount) external returns (uint);
-    function borrow(uint borrowAmount) external returns (uint);
-    function redeem(uint redeemTokens) external returns (uint);
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
-    function repayBorrow(uint repayAmount) external returns (uint);
-    function borrowBalanceCurrent(address account) external returns (uint);
-    function borrowBalanceStored(address account) external view returns (uint);
+    function mint(uint256 mintAmount) external returns (uint256);
+
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
+
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
+
+    function borrowBalanceCurrent(address account) external returns (uint256);
+
+    function borrowBalanceStored(address account) external view returns (uint256);
+
     function underlying() external view returns (address);
 }
